Splite Component test to separte component test file
